Tests/LibWeb/Text/input/wpt-import/encoding-detection/ru-ISO-8859-5.tentative.html

<!doctype html>
<title>ru ISO-8859-5</title>
<script src=../resources/testharness.js></script>
<script src=../resources/testharnessreport.js></script>
<p>Õ‚ř ‚’Š‚ ŕř‘ōŗř“ŕō Šō‹“řŘř“.</p>
<script>
setup({explicit_done:true});
onload = function() {
    test(function() {
        assert_equals(document.characterSet, "ISO-8859-5", 'Expected ISO-8859-5');
    }, "Check detection result");
      done();
};
</script>
